Move the lobby timer into its own class.

+++ src/main/java/me/limebyte/battlenight/core/tosort/PlayerData.java
@@ -43,7 +43,7 @@ public class PlayerData {
     private float flySpeed;
     private int foodLevel;
     private int gameMode;
-    private int health;
+    private double health;
     private ItemStack[] invItems;
     private ItemStack[] invArmour;
     private int level;
